4.21

+++ part4/4.1/controllers/blogs.js
@@ -35,6 +35,17 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  
+  const findBlog = await Blog.findById(request.params.id)
+  if (user.id.toString() !== findBlog.user.toString()) {
+    return response.status(401).json({ error: 'invalid user'})
+  }
+
   try {
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
